requirements.txt
torch==2.3.1
numpy==1.26.4
fastapi==0.111.0
plotly==5.22.0
pymongo==4.7.3
motor==3.4.0
pydantic-settings==2.3.3